move exon back, only callable by wrapper. move recon scripts to class
