+++ setup.cfg
@@ -66,7 +66,7 @@ console_scripts =
     circtools_primex_formatter = circtools.scripts.r_runner:circtools_primex_formatter
     circtools_primex_wrapper = circtools.scripts.r_runner:circtools_primex_wrapper
     circtools_quickcheck_wrapper = circtools.scripts.r_runner:circtools_quickcheck_wrapper
-    circtools_reconstruct_visualization = circtools.scripts.r_runner:circtools_reconstruct_visualization
+    circtools_reconstruct_visualization = circtools.reconstruct.r_runner:circtools_reconstruct_visualization
     circtools_sirna_formatter = circtools.scripts.r_runner:circtools_sirna_formatter
     circtools_reconstruct_coverage_graph = circtools.scripts.r_runner:circtools_reconstruct_coverage_graph
     circtools_reconstruct_summarized_coverage_profiles = circtools.scripts.r_runner:circtools_reconstruct_summarized_coverage_profiles
